One more warning fix.

+++ Windows/WindowsHost.cpp
@@ -88,9 +88,9 @@ WindowsHost::WindowsHost(HWND mainWindow) {
 	input.push_back(std::shared_ptr<InputDevice>(new XinputDevice()));
 	//find all connected DInput devices of class GamePad
 	size_t numDInputDevs = DinputDevice::getNumPads();
-	for (int i = 0; i < numDInputDevs; i++)
+	for (size_t i = 0; i < numDInputDevs; i++)
 	{
-		input.push_back(std::shared_ptr<InputDevice>(new DinputDevice(i)));
+		input.push_back(std::shared_ptr<InputDevice>(new DinputDevice(static_cast<int>(i))));
 	}
 	keyboard = std::shared_ptr<KeyboardDevice>(new KeyboardDevice());
 	input.push_back(keyboard);
